feat(wiki): seed Ito project wiki
