+++ .ito/wiki/index.md
@@ -1,0 +1,24 @@
+# Ito Wiki Index
+
+This wiki is an Ito-scoped knowledge layer for LLM-assisted planning,
+research, review, and archive work. It summarizes durable knowledge from Ito
+artifacts without replacing those artifacts as source truth.
+
+## Start Here
+
+- [[overview]] - high-level project and wiki usage synthesis
+- [[topics/workflow]] - change lifecycle, proposal, task, archive, and traceability flow
+- [[topics/runtime-and-storage]] - filesystem, SQLite, backend, coordination, and worktree model
+- [[topics/distribution-and-agents]] - agent instructions, skills, templates, and harness surfaces
+- [[topics/validation-quality-gates]] - validation, pre-push hooks, coverage, and repo guardrails
+- [[research/rust-port]] - durable synthesis from the Rust port research notes
+- [[log]] - notable wiki maintenance events
+- [[_meta/schema]] - page model, authority levels, and source rules
+- [[_meta/status]] - current freshness and known gaps
+
+## Source Boundary
+
+Default sources are Ito-owned artifacts: `docs/ito/specs/`,
+`docs/ito/changes/`, `.ito/research/`, `.ito/project.md`, and
+`.ito/AGENTS.md`. Link external project files only when they clarify Ito
+decisions.
