+++ .ito/wiki/log.md
@@ -1,0 +1,9 @@
+# Ito Wiki Log
+
+## 2026-05-27 Initial Seed
+
+- Created repo-local wiki scaffold and first topic pages.
+- Seeded workflow, runtime/storage, distribution/agents, quality gates, and
+  Rust port research synthesis.
+- Source window: current `docs/ito/specs/`, active/archived change summaries,
+  and existing `.ito/research/` notes.
